refactor: move ssm source test to ssm tests
