+++ tests/aws/services/events/test_events_targets.py
@@ -8,16 +8,13 @@ import pytest
 from localstack import config
 from localstack.aws.api.lambda_ import Runtime
 from localstack.testing.aws.util import is_aws_cloud
-from localstack.testing.config import TEST_AWS_ACCOUNT_ID, TEST_AWS_REGION_NAME
 from localstack.testing.pytest import markers
 from localstack.utils.aws import arns
 from localstack.utils.strings import short_uid
 from localstack.utils.sync import retry
 from localstack.utils.testutil import check_expected_lambda_log_events_length
 from tests.aws.scenario.kinesis_firehose.conftest import get_all_expected_messages_from_s3
-from tests.aws.services.events.helper_functions import (
-    sqs_collect_messages,
-)
+from tests.aws.services.events.helper_functions import sqs_collect_messages
 from tests.aws.services.events.test_events import EVENT_DETAIL, TEST_EVENT_PATTERN
 from tests.aws.services.firehose.helper_functions import get_firehose_iam_documents
 from tests.aws.services.kinesis.helper_functions import get_shard_iterator
@@ -803,71 +800,3 @@ class TestEventTargetSqs:
 
 # class TestEventTargetStepFunctions:
 #     # TODO add test for StepFunctions
-
-
-class TestEventTargetSystemsManager:
-    @markers.aws.needs_fixing  # TODO: Reason add permission and correct policies
-    @pytest.mark.parametrize("strategy", ["standard", "domain", "path"])
-    def test_trigger_event_on_systems_manager_change(
-        self, monkeypatch, aws_client, clean_up, strategy
-    ):
-        monkeypatch.setattr(config, "SQS_ENDPOINT_STRATEGY", strategy)
-
-        rule_name = "rule-{}".format(short_uid())
-        target_id = "target-{}".format(short_uid())
-
-        # create queue
-        queue_name = "queue-{}".format(short_uid())
-        queue_url = aws_client.sqs.create_queue(QueueName=queue_name)["QueueUrl"]
-        queue_arn = arns.sqs_queue_arn(queue_name, TEST_AWS_ACCOUNT_ID, TEST_AWS_REGION_NAME)
-
-        # put rule listening on SSM changes
-        ssm_prefix = "/test/local/"
-        aws_client.events.put_rule(
-            Name=rule_name,
-            EventPattern=json.dumps(
-                {
-                    "detail": {
-                        "name": [{"prefix": ssm_prefix}],
-                        "operation": [
-                            "Create",
-                            "Update",
-                            "Delete",
-                            "LabelParameterVersion",
-                        ],
-                    },
-                    "detail-type": ["Parameter Store Change"],
-                    "source": ["aws.ssm"],
-                }
-            ),
-            State="ENABLED",
-            Description="Trigger on SSM parameter changes",
-        )
-
-        # put target
-        aws_client.events.put_targets(
-            Rule=rule_name,
-            Targets=[{"Id": target_id, "Arn": queue_arn, "InputPath": "$.detail"}],
-        )
-
-        param_suffix = short_uid()
-
-        # change SSM param to trigger event
-        aws_client.ssm.put_parameter(
-            Name=f"{ssm_prefix}/test-{param_suffix}", Value="value1", Type="String"
-        )
-
-        def assert_message():
-            resp = aws_client.sqs.receive_message(QueueUrl=queue_url)
-            result = resp.get("Messages")
-            body = json.loads(result[0]["Body"])
-            assert body == {
-                "name": f"/test/local/test-{param_suffix}",
-                "operation": "Create",
-            }
-
-        # assert that message has been received
-        retry(assert_message, retries=7, sleep=0.3)
-
-        # clean up
-        clean_up(rule_name=rule_name, target_ids=target_id)
